app/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
frozenlist==1.4.0
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.37
griffe==0.36.2
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
makefile==1.1.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==0.5.0
mkdocs-material==9.4.2
mkdocs-material-extensions==1.2
mkdocstrings==0.23.0
mkdocstrings-python==1.7.0
multidict==6.0.4
numpy==1.25.2
openai==0.28.1
packaging==23.1
paginate==0.5.6
pandas==2.1.1
pathspec==0.11.2
Pillow==9.5.0
platformdirs==3.10.0
protobuf==4.24.3
pyarrow==13.0.0
pydeck==0.8.0
Pygments==2.16.1
pymdown-extensions==10.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.27.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.17.0
